CRISTIANO-RONALDO-GOALS-DATA-ANALYSIS
The "Cristiano Ronaldo Goals Data" project is a data analysis and visualization project that focuses on understanding the pattern of goals scored by Cristiano Ronaldo. The project utilizes a dataset containing information about Ronaldo's goals, including details such as the range of shots, distance of shots, and area of shots.

Through this project, we aim to uncover insights into Ronaldo's goal-scoring tendencies, such as his preferred shooting ranges, areas of the pitch where he is most effective, and the distribution of his goals by distance. By employing data analysis and visualization techniques, the project seeks to provide a comprehensive understanding of Ronaldo's goal-scoring proficiency and the factors influencing his success on the field.